+++ packages/redtun-common/src/logging/index.ts
@@ -2,12 +2,13 @@ import winston from "winston";
 
 const logger = winston.createLogger({
   level: process.env.LOG_LEVEL ?? "info",
-  format: winston.format.combine(winston.format.colorize(), winston.format.simple()),
+  format: winston.format.combine(
+    winston.format.colorize(),
+    winston.format.printf(options => {
+      return `[${options.moduleName}] ${options.level}: ${options.message}`;
+    }),
+  ),
   transports: [new winston.transports.Console()],
 });
 
-export const getLogger = (module: string) => {
-  const childLogger = logger.child({ name: module });
-  childLogger.format = winston.format.printf(({ level, message }) => `${level}: ${module}: ${message}`);
-  return childLogger;
-};
+export const getLogger = (moduleName: string) => logger.child({ moduleName });