+++ packages/redtun-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @neonredwood/redtun-common
 
+## 1.0.3
+
+### Patch Changes
+
+- logging improvements, npm packaging improvements
+
 ## 1.0.2
 
 ### Patch Changes